backend/requirements.txt
fastapi
sqlmodel
uvicorn[standard]
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
aiosmtplib
jinja2
